src/python/requirements.txt
tensorflow 
scikit-learn
scipy
numpy
pillow
pandas
icecream
elasticsearch
configparser
langchain
faiss-cpu
flask
uwsgi